requirements.txt
numpy==1.24.3
Flask==3.0.3
python-dotenv
asyncio